make reloadHiddenFieldsForInputAndFilterData public
FIX hidden override files are set in the InputAndFilterData. However, if
any of the fields are changed after the panel is drawn, then the new
values are not used during validation even if InputAndFilterData is a
set in Atdl4jOptions.

By calling reloadHiddenFieldsForInputAndFilterData before
validateStrategy, you can pass in the latest FIX values for validation
without redrawing the panel.